pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "food-combo-bot"
version = "0.1.0"
description = "Telegram-бот для консультаций по сочетанию блюд"
authors = [
    {name = "Иван Семенкин", email = "isemenkin@gmail.com"}
]

requires-python = ">=3.8"

dependencies = [
    "aiogram>=3.3.0",
    "openai>=1.90.0",
    "langchain>=0.3.0",
    "langchain-community",
    "langchain-huggingface",
    "sentence-transformers",
    "python-dotenv>=1.0.0",
    "redis>=6.0.0",
    "langchain-groq>=0.3.0",
    "python==3.8.9"
]

[tool.setuptools]
py-modules = []
